Move suitable code to elastic-job-lite-tracing-rdb
