+++ elastic-job-lite-console/src/main/java/org/apache/shardingsphere/elasticjob/lite/console/restful/EventTraceHistoryRESTfulAPI.java
@@ -22,7 +22,7 @@ import org.apache.shardingsphere.elasticjob.lite.console.domain.EventTraceDataSo
 import org.apache.shardingsphere.elasticjob.lite.console.service.EventTraceDataSourceConfigurationService;
 import org.apache.shardingsphere.elasticjob.lite.console.service.impl.EventTraceDataSourceConfigurationServiceImpl;
 import org.apache.shardingsphere.elasticjob.lite.console.util.SessionEventTraceDataSourceConfiguration;
-import org.apache.shardingsphere.elasticjob.lite.event.rdb.JobEventRdbSearch;
+import org.apache.shardingsphere.elasticjob.lite.tracing.rdb.JobEventRdbSearch;
 import org.apache.shardingsphere.elasticjob.lite.tracing.type.JobExecutionEvent;
 import org.apache.shardingsphere.elasticjob.lite.tracing.type.JobStatusTraceEvent;
 import org.apache.commons.dbcp.BasicDataSource;
