+++ elastic-job-lite-core/src/test/resources/logback-test.xml
@@ -37,7 +37,7 @@
     </root>
 
     <logger name="org.apache.curator.framework.recipes.leader.LeaderSelector" level="OFF" />
-    <logger name="org.apache.shardingsphere.elasticjob.lite.event.rdb.JobEventRdbConfiguration" level="OFF" />
+    <logger name="org.apache.shardingsphere.elasticjob.lite.tracing.rdb.JobEventRdbConfiguration" level="OFF" />
     <logger name="org.apache.shardingsphere.elasticjob.lite.tracing.JobEventBus" level="OFF" />
     <logger name="org.apache.shardingsphere.elasticjob.lite.handler.error.impl.LogJobErrorHandler" level="OFF" />
     <logger name="org.apache.shardingsphere.elasticjob.lite.internal.monitor.MonitorService" level="OFF" />
