+++ elastic-job-lite-tracing/elastic-job-lite-tracing-rdb/src/main/java/org/apache/shardingsphere/elasticjob/lite/tracing/rdb/JobEventRdbConfiguration.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,33 +15,35 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.event.rdb;
+package org.apache.shardingsphere.elasticjob.lite.tracing.rdb;
 
+import org.apache.shardingsphere.elasticjob.lite.tracing.JobEventConfiguration;
 import org.apache.shardingsphere.elasticjob.lite.tracing.JobEventListener;
-import org.apache.shardingsphere.elasticjob.lite.tracing.type.JobExecutionEvent;
-import org.apache.shardingsphere.elasticjob.lite.tracing.type.JobStatusTraceEvent;
+import org.apache.shardingsphere.elasticjob.lite.tracing.JobEventListenerConfigurationException;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
 import javax.sql.DataSource;
+import java.io.Serializable;
 import java.sql.SQLException;
 
 /**
- * Job event RDB listener.
+ * Job event RDB configuration.
  */
-public final class JobEventRdbListener extends JobEventRdbIdentity implements JobEventListener {
+@RequiredArgsConstructor
+@Getter
+public final class JobEventRdbConfiguration extends JobEventRdbIdentity implements JobEventConfiguration, Serializable {
     
-    private final JobEventRdbStorage repository;
+    private static final long serialVersionUID = 3344410699286435226L;
     
-    public JobEventRdbListener(final DataSource dataSource) throws SQLException {
-        repository = new JobEventRdbStorage(dataSource);
-    }
-    
-    @Override
-    public void listen(final JobExecutionEvent executionEvent) {
-        repository.addJobExecutionEvent(executionEvent);
-    }
+    private final transient DataSource dataSource;
     
     @Override
-    public void listen(final JobStatusTraceEvent jobStatusTraceEvent) {
-        repository.addJobStatusTraceEvent(jobStatusTraceEvent);
+    public JobEventListener createJobEventListener() throws JobEventListenerConfigurationException {
+        try {
+            return new JobEventRdbListener(dataSource);
+        } catch (final SQLException ex) {
+            throw new JobEventListenerConfigurationException(ex);
+        }
     }
 }
