+++ elastic-job-lite-tracing/elastic-job-lite-tracing-rdb/src/test/java/org/apache/shardingsphere/elasticjob/lite/tracing/rdb/JobEventRdbSearchTest.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,15 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.event.rdb;
+package org.apache.shardingsphere.elasticjob.lite.tracing.rdb;
 
 import org.apache.commons.dbcp.BasicDataSource;
-import org.apache.shardingsphere.elasticjob.lite.context.ExecutionType;
 import org.apache.shardingsphere.elasticjob.lite.tracing.type.JobExecutionEvent;
 import org.apache.shardingsphere.elasticjob.lite.tracing.type.JobStatusTraceEvent;
 import org.apache.shardingsphere.elasticjob.lite.tracing.type.JobStatusTraceEvent.Source;
 import org.apache.shardingsphere.elasticjob.lite.tracing.type.JobStatusTraceEvent.State;
-import org.apache.shardingsphere.elasticjob.lite.util.env.IpUtils;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
@@ -35,7 +33,7 @@ import java.util.Map;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
-public class JobEventRdbSearchTest {
+public final class JobEventRdbSearchTest {
     
     private static JobEventRdbStorage storage;
     
@@ -55,14 +53,14 @@ public class JobEventRdbSearchTest {
     
     private static void initStorage() {
         for (int i = 1; i <= 500; i++) {
-            JobExecutionEvent startEvent = new JobExecutionEvent(IpUtils.getHostName(), IpUtils.getIp(), "fake_task_id", "test_job_" + i, JobExecutionEvent.ExecutionSource.NORMAL_TRIGGER, 0);
+            JobExecutionEvent startEvent = new JobExecutionEvent("localhost", "127.0.0.1", "fake_task_id", "test_job_" + i, JobExecutionEvent.ExecutionSource.NORMAL_TRIGGER, 0);
             storage.addJobExecutionEvent(startEvent);
             if (i % 2 == 0) {
                 JobExecutionEvent successEvent = startEvent.executionSuccess();
                 storage.addJobExecutionEvent(successEvent);
             }
             storage.addJobStatusTraceEvent(new JobStatusTraceEvent(
-                    "test_job_" + i, "fake_failed_failover_task_id", "fake_slave_id", Source.LITE_EXECUTOR, ExecutionType.FAILOVER.name(), "0", State.TASK_FAILED, "message is empty."));
+                    "test_job_" + i, "fake_failed_failover_task_id", "fake_slave_id", Source.LITE_EXECUTOR, "FAILOVER", "0", State.TASK_FAILED, "message is empty."));
         }
     }
